% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appTEKRABber.R
\name{appTEKRABber}
\alias{appTEKRABber}
\title{Visualize TEKRABber results with shiny app}
\usage{
appTEKRABber(DEresult, corrRef, corrCompare, metadata)
}
\arguments{
\item{DEresult}{the output variable from using DEgeneTE()}

\item{corrRef}{the correlation result of your reference species using corrOthologTE()}

\item{corrCompare}{the correlation result of your compare species using corrOrthologTE()}

\item{metadata}{the same metadata you use for DEgeneTE()}
}
\value{
an app can display differentially expressed genes/TE and the correlation results
}
\description{
To help user explore their results using TEKRABber, it visualize the results
using a self-written shiny app with two tabs, including the expression and correlation of
genes and TEs.
}
\examples{
data(ctInputDE)
geneInputDE <- ctInputDE$gene
teInputDE <- ctInputDE$te

metaExp <- data.frame(experiment = c(rep("control", 5), rep("treatment", 5)))
rownames(metaExp) <- colnames(geneInputDE)
metaExp$experiment <- factor(metaExp$experiment, levels = c("control", "treatment"))

resultDE <- DEgeneTE(
  geneTable = geneInputDE,
  teTable = teInputDE,
  metadata = metaExp,
  contrastVector = c("experiment", "control", "treatment"),
  expDesign = FALSE
)

#library(SummarizedExperiment)
#data(ctCorr)
#geneConCorrInput <- assay(ctCorr$geneCorr[, ctCorr$geneCorr$experiment=="control"])
#teConCorrInput <- assay(ctCorr$teCorr[, ctCorr$teCorr$experiment=="control"])
#geneTreatCorrInput <- assay(ctCorr$geneCorr[, ctCorr$geneCorr$experiment=="treatment"])
#teTreatCorrInput <- assay(ctCorr$teCorr[, ctCorr$teCorr$experiment=="treatment"])

#controlCorr <- corrOrthologTE(
#  geneInput = geneConCorrInput,
#  teInput = teConCorrInput,
#  corrMethod = "pearson",
#  padjMethod = "fdr",
#  filename = "controlCorrResult.csv"
#)

#treatmentCorr <- corrOrthologTE(
#  geneInput = geneTreatCorrInput,
#  teInput = teTreatCorrInput,
#  corrMethod = "pearson",
#  padjMethod = "fdr",
#  filename = "treatmentCorrResult.csv"
#)

#appTEKRABber(
#  DEresult = resultDE,
#  corrRef = controlCorr,
#  corrCompare = treatmentCorr,
#  metadata = metaExp
#)

}
