% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appTEKRABber2.R
\name{appTEKRABber2}
\alias{appTEKRABber2}
\title{appTEKRABber2}
\usage{
appTEKRABber2(corrRef, corrCompare = NULL, DEobject)
}
\arguments{
\item{corrRef}{correlation results for reference using corrOrtholgScale()}

\item{corrCompare}{correlation results for comparison using 
corrOrthologScale()}

\item{DEobject}{DE object using DEgeneTE()}
}
\value{
provide an interactive shinyapp
}
\description{
appTEKRABber2
}
\examples{



}
