% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assay_tekcorrset.R
\name{assay_tekcorrset}
\alias{assay_tekcorrset}
\title{Access gene and transposable elements expression data}
\usage{
assay_tekcorrset(tecorrset, expType, sample)
}
\arguments{
\item{tecorrset}{TekCorrSet object}

\item{expType}{Indicate which data you want to access. It should be "gene" or "te".}

\item{sample}{The species name or experimental design. It should be "human" and 
"chimpanzee" when you are running comparing species design. "control" and "treatment"
 are for running same species design.}
}
\value{
a dataframe contains expression genes or transposable element.
}
\description{
Access expression data from the TekCorrSet class for running vignette
demonstration.
}
\examples{
data(speciesCorr)
hmGeneCorrInput <- assay_tekcorrset(speciesCorr, "gene", "human")
hmTECorrInput <- assay_tekcorrset(speciesCorr, "te", "human")
chimpGeneCorrInput <- assay_tekcorrset(speciesCorr, "gene", "chimpanzee")
chimpTECorrInput <- assay_tekcorrset(speciesCorr, "te", "chimpanzee")
}
