% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chimpGeneCorrInput}
\alias{chimpGeneCorrInput}
\title{Chimpanzee gene dataset for running correlation analysis}
\format{
A data frame with 50 rows and 10 variables:
\describe{
  \item{SRR8750637}{sample ID}
  \item{SRR8750638}{sample ID}
  \item{SRR8750639}{sample ID}
  \item{SRR8750640}{sample ID}
  \item{SRR8750641}{sample ID}
  \item{SRR8750642}{sample ID}
  \item{SRR8750643}{sample ID}
  \item{SRR8750644}{sample ID}
  \item{SRR8750645}{sample ID}
  \item{SRR8750646}{sample ID}
}
}
\usage{
chimpGeneCorrInput
}
\description{
A normalized gene expression dataset containing 10 chimpanzee samples to perform 
a quick demonstration using TEKRABber for correlation analysis.
}
\keyword{datasets}
