% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chimpTE}
\alias{chimpTE}
\title{Transposable elements expression dataset from chimpanzee brain RNA-seq data}
\format{
A data frame with 1260 rows and 11 variables:
\describe{
  \item{teName}{Name of transposable elements}
  \item{SRR8750637}{sample ID}
  \item{SRR8750638}{sample ID}
  \item{SRR8750639}{sample ID}
  \item{SRR8750640}{sample ID}
  \item{SRR8750641}{sample ID}
  \item{SRR8750642}{sample ID}
  \item{SRR8750643}{sample ID}
  \item{SRR8750644}{sample ID}
  \item{SRR8750645}{sample ID}
  \item{SRR8750646}{sample ID}
}
}
\usage{
chimpTE
}
\description{
A dataset containing randomly slected 10 chimpanzee brain cells bulked 
RNA-seq transposable elements expression data. It is generated using in-house
script by the author using fastq data provided from (Khrameeva E et al., 2020).
}
\keyword{datasets}
