% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctInputDE}
\alias{ctInputDE}
\title{Input expression data of gene/TE for differentially expressed analysis within same species}
\format{
A data list with 2 expression counts:
\describe{
  \item{gene}{input gene data for DE analysis comparing control and treatment}
  \item{te}{input TE data for DE analysis comparing control and treatment}
}
}
\usage{
ctInputDE
}
\description{
Dataset contains 2 expression toy data for demonstrating how to use TEKRABber
on experimental design within the case of comparing same species.
}
\keyword{datasets}
