% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fetchDataHmChimp}
\alias{fetchDataHmChimp}
\title{Example output comparing human and chimpanzee data using orhtologScale()}
\format{
An object contains 2 elements:
\describe{
    \item{\strong{orthologTable}}{orthology information from Ensembl}
    \item{\strong{scaleFactor}}{scaling factor to normalize data}
}
}
\usage{
data(fetchDataHmChimp)
}
\description{
An output list of data contains 7 elements after using orthologScale(), 
including (1) orthology table comparing human and chimpanzee. (2) scaling 
factor for orthologous genes (3) gene count table from reference species (4)
 gnee count table from species you want to compare (5) scaling factor for TEs
(6) TE count table from reference species (7) TE count table from the 
species you want to compare. The aim to provide this dataset is to save 
time for user running the vignettes and give a 
template for demonstration.
}
\examples{
data(fetchDataHmChimp)
fetchData <- fetchDataHmChimp
fetchData$orthologTable
fetchData$scaleFactor
}
\keyword{datasets}
