% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneInputDE}
\alias{geneInputDE}
\title{Gene expression dataset for running differentially expressed analysis (DE)}
\format{
A data frame with 33223 rows and 10 variables:
\describe{
  \item{control_1}{control sample ID}
  \item{control_2}{control sample ID}
  \item{control_3}{control sample ID}
  \item{control_4}{control sample ID}
  \item{control_5}{control sample ID}
  \item{treatment_6}{treatment sample ID}
  \item{treatment_7}{treatment sample ID}
  \item{treatment_8}{treatment sample ID}
  \item{treatment_9}{treatment sample ID}
  \item{treatment_10}{treatment sample ID}
}
}
\usage{
geneInputDE
}
\description{
A toy gene expression dataset containing 5 controls and 5 treatments to demonstrate
the case if user want to use TEKRABber for DE analysis in the same species.
}
\keyword{datasets}
