% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg38_panTro6_rmsk}
\alias{hg38_panTro6_rmsk}
\title{Repeatmasker track annotations with human and chimpanzee}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 12550 rows and 4 columns.
}
\usage{
data(hg38_panTro6_rmsk)
}
\description{
This Repeatmasker track annotations table was first downloaded from UCSC 
Genome Table Browser and it included the name, class, and average gene 
length in repeats(transposable elements). This data is used for demonstrate 
an example for user how to provide a annotation table to normalize their 
data which in this case comparing human(hg38) to chimpanzee(panTro6).
}
\examples{
data(hg38_panTro6_rmsk)
}
\keyword{datasets}
