% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmGene}
\alias{hmGene}
\title{Gene expression dataset from human brain RNA-seq data}
\format{
A data frame with 64252 rows and 11 variables:
\describe{
  \item{Geneid}{Ensembl gene ID}
  \item{SRR8750453}{sample ID}
  \item{SRR8750454}{sample ID}
  \item{SRR8750455}{sample ID}
  \item{SRR8750456}{sample ID}
  \item{SRR8750457}{sample ID}
  \item{SRR8750458}{sample ID}
  \item{SRR8750459}{sample ID}
  \item{SRR8750460}{sample ID}
  \item{SRR8750461}{sample ID}
  \item{SRR8750462}{sample ID}
}
}
\usage{
hmGene
}
\description{
A dataset containing randomly slected 10 human brain cells bulked RNA-seq
gene expression data. It is generated using in-house script by the author 
using fastq data provided from (Khrameeva E et al., 2020).
}
\keyword{datasets}
