% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmGeneCorrInput}
\alias{hmGeneCorrInput}
\title{Human gene dataset for running correlation analysis}
\format{
A data frame with 50 rows and 10 variables:
\describe{
  \item{SRR8750453}{sample ID}
  \item{SRR8750454}{sample ID}
  \item{SRR8750455}{sample ID}
  \item{SRR8750456}{sample ID}
  \item{SRR8750457}{sample ID}
  \item{SRR8750458}{sample ID}
  \item{SRR8750459}{sample ID}
  \item{SRR8750460}{sample ID}
  \item{SRR8750461}{sample ID}
  \item{SRR8750462}{sample ID}  
}
}
\usage{
hmGeneCorrInput
}
\description{
A normalized gene expression dataset containing 10 human samples to perform 
a quick demonstration using TEKRABber for correlation analysis.
}
\keyword{datasets}
