% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmTE}
\alias{hmTE}
\title{Transposable elements expression dataset from human brain RNA-seq data}
\format{
A data frame with 992 rows and 11 variables:
\describe{
  \item{teName}{Name of transposable elements}
  \item{SRR8750453}{sample ID}
  \item{SRR8750454}{sample ID}
  \item{SRR8750455}{sample ID}
  \item{SRR8750456}{sample ID}
  \item{SRR8750457}{sample ID}
  \item{SRR8750458}{sample ID}
  \item{SRR8750459}{sample ID}
  \item{SRR8750460}{sample ID}
  \item{SRR8750461}{sample ID}
  \item{SRR8750462}{sample ID}
}
}
\usage{
hmTE
}
\description{
A dataset containing randomly slected 10 human brain cells bulked RNA-seq
transposable elements expression data. It is generated using in-house script 
by the author using fastq data provided from (Khrameeva E et al., 2020).
}
\keyword{datasets}
