% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_corr}
\alias{rcpp_corr}
\title{Estimate the correlation between genes and transposable elements}
\usage{
rcpp_corr(df1, df2, Method)
}
\arguments{
\item{df1}{First dataframe}

\item{df2}{Second dataframe}

\item{Method}{correlation method}
}
\value{
a dataframe containing correlation results
}
\description{
Estimate the correlation between genes and transposable elements
}
