% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speciesCorr}
\alias{speciesCorr}
\title{A subsets of normalized Gene/TE expression data from human/chimpanzee brain RNA-seq for 
correlation analysis demonstration}
\format{
An object of class "TekCorrSet" which contains 4 expression counts 
and you can access it specifying the parameters using assay_tekcorrset():
\describe{
  \item{assay_tekcorrset(speciesCorr, "gene", "human")}{human gene expression data}
  \item{assay_tekcorrset(speciesCorr, "te", "human")}{human TE expression data}
  \item{assay_tekcorrset(speciesCorr, "gene", "chimpanzee")}{chimpanzee gene expression data}
  \item{assay_tekcorrset(speciesCorr, "te", "chimpanzee")}{chimpanzee TE expression data}
}
}
\usage{
data(speciesCorr)
hmGeneCorrInput <- assay_tekcorrset(speciesCorr, "gene", "human")
hmTECorrInput <- assay_tekcorrset(speciesCorr, "te", "human")
chimpGeneCorrInput <- assay_tekcorrset(speciesCorr, "gene", "chimpanzee")
chimpTECorrInput <- assay_tekcorrset(speciesCorr, "te", "chimpanzee")
}
\description{
An object of class "TekCorrSet" which contains 4 expression counts. 
These data are generated from speciesCounts using TEKRABber. For a quick demo, 
we only select 50 orthologs and 50 transposable elements.
}
\keyword{datasets}
