% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teConCorrInput}
\alias{teConCorrInput}
\title{Transposable elements expression data in control group for running correlation analysis}
\format{
A data frame with 10 rows and 5 variables:
\describe{
  \item{control_1}{control sample ID}
  \item{control_2}{control sample ID}
  \item{control_3}{control sample ID}
  \item{control_4}{control sample ID}
  \item{control_5}{control sample ID}
}
}
\usage{
teConCorrInput
}
\description{
A small toy transposable elements expression dataset containing 5 controls to
perform quick demonstration using TEKRABber for correlation analysis.
}
\keyword{datasets}
