% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teTreatCorrInput}
\alias{teTreatCorrInput}
\title{Transposable elements expression data in treatment group for running 
correlation analysis}
\format{
A data frame with 10 rows and 5 variables:
\describe{
  \item{treatment_6}{control sample ID}
  \item{treatment_7}{control sample ID}
  \item{treatment_8}{control sample ID}
  \item{treatment_9}{control sample ID}
  \item{treatment_10}{control sample ID}
}
}
\usage{
teTreatCorrInput
}
\description{
A small toy transposable elements expression dataset containing 5 treatments 
to perform quick demonstration using TEKRABber for correlation analysis.
}
\keyword{datasets}
