% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appTEKRABber.R
\name{appTEKRABber}
\alias{appTEKRABber}
\title{Visualize TEKRABber results with shiny app}
\usage{
appTEKRABber(DEresult, corrRef, corrCompare, metadata)
}
\arguments{
\item{DEresult}{the output variable from using DEgeneTE()}

\item{corrRef}{the correlation result of your reference species using corrOthologTE()}

\item{corrCompare}{the correlation result of your compare species using corrOrthologTE()}

\item{metadata}{the same metadata you use for DEgeneTE()}
}
\value{
an app can display differentially expressed genes/TE and the correlation results
}
\description{
To help user explore their results using TEKRABber, it visualize the results
using a self-written shiny app with two tabs, including the expression and correlation of
genes and TEs.
}
