% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appTEKRABber.R
\name{appTEKRABber}
\alias{appTEKRABber}
\title{appTEKRABber}
\usage{
appTEKRABber(corrRef, corrCompare, DEobject)
}
\arguments{
\item{corrRef}{correlation results for reference using corrOrtholgScale()}

\item{corrCompare}{correlation results for comparison using 
corrOrthologScale()}

\item{DEobject}{DE object using DEgeneTE()}
}
\value{
provide an interactive shinyapp
}
\description{
appTEKRABber
}
