% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrOrthologTE.R
\name{corrOrthologTE}
\alias{corrOrthologTE}
\title{Estimate correlation comparing orthologs and TEs}
\usage{
corrOrthologTE(geneInput, teInput, corrMethod = "pearson", padjMethod = "fdr", fileName)
}
\arguments{
\item{geneInput}{gene count input for correlation from using DECorrInputs()}

\item{teInput}{te count input for correlation from using DECorrInputs()}

\item{corrMethod}{correlation method, inlcuding pearson, kendall, spearman. Default is pearson}

\item{padjMethod}{method to return p-values adjusted, and default is fdr. See ?p.adjust}

\item{fileDir}{the name and path of directory for saving output files. Default is NULL.}

\item{fileName}{the name for saving outputfiles. Default is "TEKRABber_geneTECorrResult.csv"}
}
\value{
a dataframe includes Pearson's correlation coefficient, pvalue, padj
}
\description{
To estimate correlation comparing each ortholog to each TE from inputs.
You can sepcify the correlation and adjusted p-value methods (see details in parameters).
It will also generate a output csv file in ./results directory.
a output csv file with the correlation result.
}
\examples{
library(SummarizedExperiment)
data(speciesCorr)
hmGeneCorrInput <- assay_tekcorrset(speciesCorr, "gene", "human")
hmTECorrInput <- assay_tekcorrset(speciesCorr, "te", "human")

\donttest{
corrOrthologTE(
    geneInput=hmGeneCorrInput,
    teInput=hmTECorrInput,
    corrMethod="pearson",
    padjMethod="fdr",
    fileDir="."
    fileName="correlationResult.csv"
)}
}
