% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctCorr}
\alias{ctCorr}
\title{Normalized Gene/TE expression toy data in control and treatment in same species for 
correlation analysis}
\format{
A data list with 4 expression counts:
\describe{
  \item{geneCorr}{containing gene expression data from control and treatment}
  \item{teCorr}{containing TE expression data from control and treatment}
}
}
\usage{
ctCorr
}
\description{
Dataset contains 4 normalized toy data. For a quick demo, we only use a subset of rows.
}
\keyword{datasets}
