% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DECorrInputs.R
\name{DECorrInputs}
\alias{DECorrInputs}
\title{Generate all the input files for TEKRABber downstream analysis}
\usage{
DECorrInputs(orthologTable, scaleFactor, geneCountRef, 
geneCountCompare, teCountRef, teCountCompare)
}
\arguments{
\item{orthologTable}{orthologTable output from using orthologScale()}

\item{scaleFactor}{scaleFactor output from using orthologScale()}

\item{geneCountRef}{Gene counts from your reference species. First 
column should be Ensmebl gene ID.}

\item{geneCountCompare}{Gene counts from the species you want to compare. 
First column should also be Ensembl gene ID.}

\item{teCountRef}{TE counts from your reference species. First column 
should be TE's name.}

\item{teCountCompare}{TE counts from the species you want to compare. 
First column should also be TE's name.}
}
\value{
create inputs for DE analysis and correlations: 
(1) geneInputDESeq2 (2) teInputDESeq2 (3) geneCorrInputRef 
(4) geneCorrInputCompare (5) TECorrInputRef (6) TECorrInputCompare
}
\description{
Generate all the inputs files for differentially expressed 
genes/TEs analysis, and for correlation analysis. The output 
is a list containing 6 dataframes.
}
\examples{
data(speciesCounts)
hmGene <- speciesCounts$hmGene
chimpGene <- speciesCounts$chimpGene
hmTE <- speciesCounts$hmTE
chimpTE <- speciesCounts$chimpTE

## For demonstration, here we only select 1000 rows to save time
set.seed(1234)
hmGeneSample <- hmGene[sample(nrow(hmGene), 1000), ]
chimpGeneSample <- chimpGene[sample(nrow(chimpGene), 1000), ]

fetchData <- orthologScale(
    speciesRef = "hsapiens",
    speciesCompare = "ptroglodytes",
    geneCountRef = hmGeneSample,
    geneCountCompare = chimpGeneSample
)

inputBundle <- DECorrInputs(
    orthologTable=fetchData$orthologTable,
    scaleFactor=fetchData$scaleFactor,
    geneCountRef=hmGene,
    geneCountCompare=chimpGene,
    teCountRef=hmTE,
    teCountCompare=chimpTE
)
}
