% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEgeneTE.R
\name{DEgeneTE}
\alias{DEgeneTE}
\title{Estimate differentially expressed genes and TEs}
\usage{
DEgeneTE(geneTable, teTable, metadata, contrastVector, expDesign = TRUE)
}
\arguments{
\item{geneTable}{gene input table from using DECorrInputs()}

\item{teTable}{TE input table from using DECorrInputs()}

\item{metadata}{a one column dataframe with rownames same as the column name of gene/te count table. Column name must be species or experiment.}

\item{contrastVector}{your experiment design, i.e. c("species", "human", "chimpanzee")}

\item{expDesign}{Logic value for comparing between or within species. TRUE for comparing between two species, and False for comparing between control and treatment.}
}
\value{
output DESeq2 res and normalized gene counts result in ./results directory
}
\description{
To estimate differentially expressed genes and TEs, DEgeneTE() takes
gene input and TE input from the result using DECorrInputs function and return results
in ./results directory which is automatically generate by this function. You need to
specify metadata, contrastVector, and expDesign. See details in example.
}
\examples{
# if you are comparing between species:
# (1) you need to specify the column name to species in the metadata
# (2) you can use the output from DECorrInputs()
# (3) you need to set "expDesign = TRUE" (see vignettes/TEKRABber.Rmd for details)
data(geneInputDE)
data(teInputDE)
metaExp <- data.frame(experiment = c(rep("control", 5), rep("treatment", 5)))
rownames(metaExp) <- colnames(geneInputDE)
metaExp$experiment <- factor(metaExp$experiment, levels = c("control", "treatment"))

resultDE <- DEgeneTE(
  geneTable = geneInputDE,
  teTable = teInputDE,
  metadata = metaExp,
  contrastVector = c("experiment", "control", "treatment"),
  expDesign = FALSE
)
}
