% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fetchDataHmChimp}
\alias{fetchDataHmChimp}
\title{Orthology information of human and chimpanzee with scaling factor}
\format{
An object contains 2 elements:
\describe{
  \item{\strong{orthologTable}}{containing orthology information retrieving from Ensembl}
  \item{\strong{scaleFactor}}{containing the scaling factor to normalize data}
}
}
\usage{
data(fetchDataHmChimp)
fetchData <- fetchDataHmChimp
fetchData$orthologTable
fetchData$scaleFactor
}
\description{
An output list of data contains 2 elements after using orthologScale(). 
The first one is the orthology table comparing human and chimpanzee. The second one
is the scaling factor. The purpose of providing this dataset is to 
save time for user running the tutorial and give a template for demonstration.
}
\keyword{datasets}
