% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fetchDataHmChimp}
\alias{fetchDataHmChimp}
\title{Orthology information of human and chimpanzee with scaling factor}
\format{
A data list with 2 elements:
\describe{
  \item{orthologTable}{containing orthology information retrieving from Ensembl}
  \item{scaleFactor}{containing the scaling factor to normalize data}
}
}
\usage{
fetchDataHmChimp
}
\description{
A output list of data contains 2 elements after using orthologScale(). 
The first one is the orthology table comparing human and chimpanzee. The second
is the scaling factor. The purpose of preparing this datasets for user is to 
save time on running the tutorial and give a template for demonstration.
}
\keyword{datasets}
