% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthologScale.R
\name{orthologScale}
\alias{orthologScale}
\title{Get orthology information from Ensembl}
\usage{
orthologScale(speciesRef, speciesCompare, geneCountRef, 
geneCountCompare)
}
\arguments{
\item{speciesRef}{The scientific name for your reference species. 
i.e., hsapiens}

\item{speciesCompare}{The scientific name for your species to compare. 
i.e., ptroglodytes}

\item{geneCountRef}{Gene count from your reference species. First column 
should be Ensmebl gene ID}

\item{geneCountCompare}{Gene count from the species you want to compare. 
First column should also be Ensembl gene ID}
}
\value{
There are two outputs:(1) orthologTable: orthology information 
from BioMart (2) scale_factor: for normalizing expression counts
}
\description{
Get orthology information from Ensembl using biomaRt and
calculate scaling factor between two species using the confidence of
orthology score and expression counts.
}
\examples{
data(speciesCounts)
hmGene <- speciesCounts$hmGene
chimpGene <- speciesCounts$chimpGene

## For demonstration, here we only select 1000 rows to save time
set.seed(1234)
hmGeneSample <- hmGene[sample(nrow(hmGene), 1000), ]
chimpGeneSample <- chimpGene[sample(nrow(chimpGene), 1000), ]

fetchData <- orthologScale(
    speciesRef = "hsapiens",
    speciesCompare = "ptroglodytes",
    geneCountRef = hmGeneSample,
    geneCountCompare = chimpGeneSample
)
}
