% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speciesCorr}
\alias{speciesCorr}
\title{Normalized Gene/TE expression data from human/chimpanzee brain RNA-seq for 
correlation analysis.}
\format{
A data list with 2 expression counts:
\describe{
  \item{geneCorr}{containing gene expression data from human and chimpanzee}
  \item{teCorr}{containing TE expression data from human and chimpanzee}
}
}
\usage{
speciesCorr
}
\description{
Datasets containing 2 normalized data from human and chimpanzee RNA-seq 
data. For a quick demo, here we only subset a
}
\keyword{datasets}
