% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speciesCounts}
\alias{speciesCounts}
\title{Gene/TE expression data from human/chimpanzee brain RNA-seq}
\format{
A data list with 4 expression counts:
\describe{
  \item{hmGene}{human gene expression from RNA-seq}
  \item{hmTE}{human TE expression from RNA-seq}
  \item{chimpGene}{chimpanzee gene expression from RNA-seq}
  \item{chimpTE}{chimpanzee TE expression from RNA-seq}
}
}
\usage{
speciesCounts
}
\description{
Datasets contain 4 expression data from human and chimpanzee RNA-seq 
data. It is generated using in-house script using fastq data provided from
(Khrameeva E et al., 2020).
}
\keyword{datasets}
